fix(security): restore API rate limiting in the active proxy
The rate limiter lived in the root middleware.ts, which Next 16 ignores in this src/ project, so it never ran; port it into src/proxy.ts and remove the dead file.
